requirements.txt
fastapi
pydantic
uvicorn[standard]
google-cloud-aiplatform[adk,agent_engines]
google-adk